Fitting Linear Regression Models

`lr()` serves the purpose of fitting linear regression models. It can be used for linear regression, determination of confidence intervals for relevant parameters, computation of \(R^2\), and facilitation of the F Test.

lr(Y, X, CI = TRUE, CI.level = 0.95, Rcpp = FALSE)

Arguments

Y

a numeric vector, treated as outcome.

X

a numerical matrix, treated as predictors.

CI

logical, "TRUE" by default. If "TRUE", the results will output the confidence interval of betas.

CI.level

a numeric number specifying the alpha-level when calculate CIs of betas (0.95 by default). Only use when "CI.beta = TRUE"

Rcpp

"FALSE" by default. If "TRUE", the function will use cpp code to speed up the computation.

Value

A list contains statistics from the regression model

Examples

 # fit a regression model
 lr(state.x77[, 4], state.x77[, c(1:3)], Rcpp = FALSE)
#> $Coefficents
#>                 Estimate   Std. Error    t value     Pr(>|t|)
#>             7.120230e+01 1.494254e+00 47.6507505 0.0000000000
#> Population -1.024175e-05 3.734214e-05 -0.2742680 0.7851066424
#> Income      2.477497e-04 2.998995e-04  0.8261089 0.4130066381
#> Illiteracy -1.178787e+00 2.974291e-01 -3.9632554 0.0002556016
#> 
#> $Confidence_Interval
#>                    Lower         Upper
#>             6.819453e+01  7.421008e+01
#> Population -8.540757e-05  6.492407e-05
#> Income     -3.559167e-04  8.514161e-04
#> Illiteracy -1.777481e+00 -5.800938e-01
#> 
#> $R_squared
#>      Multiple_R_squared Adjusted_R_squared
#> [1,]          0.3558849          0.3138774
#> 
#> $F_statistics
#>      F statistics      p value df.ssr df.sse
#> [1,]     8.471935 0.0001372993      3     46
#>