Example Data: state.x77

state.x77 related to the 50 states of the United States of America.

state.x77

Format

A dataframe with 50 rows and 8 numerical columns giving the following statistics in the respective columns.

Population:

population estimate as of July 1,1975

Income:

per capita income (1974)

Illiteracy:

illiteracy (1970, percent of population)

Life Exp:

life expectancy in years (1969--71)

Murder:

murder and non-negligent manslaughter rate per 100,000 population (1976)

HS Grad:

percent high-school graduates (1970)

Frost:

mean number of days with minimum temperature below freezing (1931--1960) in capital or large city

Area:

land area in square miles